+++ deploy/handlers/HCurveDaoNoMint.js
@@ -1,0 +1,7 @@
+const { deploy } = require('../utils/deploy.js');
+
+module.exports = async () => {
+  await deploy('HCurveDaoNoMint');
+};
+
+module.exports.tags = ['HCurveDaoNoMint'];